TST: Moved the FakePV and wrappers into `ophyd.utils` for external use
